Almost finished user

+++ service/api/delete-deleteUser.go
@@ -1,6 +1,7 @@
 package api
 
 import (
+	"database/sql"
 	"net/http"
 	"strconv"
 
@@ -14,26 +15,22 @@ func (rt *_router) deleteUser(w http.ResponseWriter, r *http.Request, ps httprou
 	w.Header().Set("Content-type", "application/json")
 
 	// Gets the user from the request body
-	userID, err := strconv.Atoi(ps.ByName("user_id"))
+	userID, err := strconv.Atoi(ps.ByName("id"))
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
 
-	ctx.Logger.WithField("id", userID).Debug("deleting user")
-
 	// Checks if user exists
-	// userExists, err := rt.db.ExistsName(user.Username)
-	// if err != nil {
-	// 	http.Error(w, err.Error(), http.StatusInternalServerError)
-	// 	return
-	// }
-
-	// If user does not exist, return 404
-	// if !userExists {
-	// 	http.Error(w, "User not found", http.StatusNotFound)
-	// 	return
-	// }
+	_, err = rt.db.GetUserById(userID)
+	if err == sql.ErrNoRows {
+		http.Error(w, "User not found", http.StatusNotFound)
+		return
+	}
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
 
 	// If user exists, delete user
 	err = rt.db.DeleteUser(userID)
